refactor: normalize UserDropdown component
