--- web/components/common/UserDropdown/index.ts
@@ -1,1 +1,0 @@
-export { default } from './UserDropdown';
