--- web/components/common/index.ts
@@ -1,1 +1,0 @@
-export { default as UserDropdown } from './UserDropdown';
