+++ web/components/ui/Header/Header.tsx
@@ -1,5 +1,5 @@
 import { Layout, Tag, Tooltip } from 'antd';
-import { UserDropdown } from '../../common';
+import { UserDropdown } from '../../common/UserDropdown/UserDropdown';
 import { OwncastLogo } from '../../common/OwncastLogo/OwncastLogo';
 import s from './Header.module.scss';
